fixing tings

+++ libft.h
@@ -9,7 +9,7 @@ int				ft_isalnum(int	c);
 int				ft_isdigit(int	c);
 int				ft_isascii(int	c);
 int				ft_isprint(int	c);
-int				ft_strlen(char const *str);
+size_t			ft_strlen(char const *str);
 void			*ft_memset(void *dest, int	c, size_t	len);
 void			*ft_memmove(void	*dst, const void	*src, size_t	len);
 void			ft_bzero(void	*s, size_t	n);
